create child process mostly from rust side
This change attempts to move most of the work to the rust side of things
by relying heavily on jni. Part of this change includes cleaning unused
code on both the java and rust side. This is done to try and make the
work that was done easy to follow and understand.