create multiple isolated process services
To do this we need a specific class for each service. This makes it so
that we are capped on the number of services that we can create at
once. The boilerplate is a little annoying but this is how chrome deals
with the mutliple services so seems like the "correct" way to do it.